// src/app/api/models/page-bc-recovery-priorities.ts
/* tslint:disable */
/* eslint-disable */
import { BcRecoveryPriorities } from './bc-recovery-priorities';
import { PageableObject } from './pageable-object';
import { SortObject } from './sort-object';
export interface PageBcRecoveryPriorities {
  content?: Array<BcRecoveryPriorities>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  pageable?: PageableObject;
  size?: number;
  sort?: SortObject;
  totalElements?: number;
  totalPages?: number;
}